refactor: move spell checking into worker
## cspell-eslint-plugin

Move the logic out of `cspell-eslint-plugin.ts` and into its own file.

The plan is to make an async version, but it needs to be in its own file and cannot call into eslint context.
